requirements.txt
msal==1.30.0
requests==2.32.3
typer==0.19.2
rich>=10.11.0,<13.0.0
pytest==8.3.3
pytest-mock==3.14.0
python-dotenv==1.0.0
